PSR4 Fix + Package Fix
I have updated all of the following files for reasons listed for each of them. Overall just PSR4 fix.

.gitignore:
 + Added *.ignore.* to ignore files for local dev

src\GitOdin_load.php:
 + Added this for Users not using PSR4 and Composer for dep mgr.

composer.json:
 ~ Updated meta author tag

example\example.php:
 ~ Fix missed php mistakes
 ++
example\example_auth.php:
examples\php_cli.php:
examples\php_cli_load.php:
 ~ Updated require statments
 ~ Updated the Namespaces for `use` statments for PSR4

src\GitOdin.php:
 ~ Updated Comments for Debug and Datatypes
 ~ Fix for the Sending Payload
 ++
src\Request\Authentication.php:
 ~ Can't use boolean as datatype, use bool
 ++
src\Request\Event.php
src\Request\EventGroup.php
src\Request\Payload.php
src\Traits\InstanceInterface.php
 - Removed require_once Statments
 ~ Updated the Namespaces for `use` statments for PSR4
