+++ .gitignore
@@ -1,2 +1,3 @@
 config.json
 /vendor/
+*.ignore.*
