+++ src/GitOdin_load.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * This is to be used when you are NOT using composer where PSR4 is used to
+ *  dynamically load the files when needed.
+ *
+ * This will load all files for this package.
+ */
+
+require_once("Request/Payload.php");
+require_once("Request/Authentication.php");
+require_once("Request/Event.php");
+require_once("Request/EventGroup.php");
+require_once("Traits/InstanceInstance.php");
+require_once("GitOdin.php");
+
+?>
